% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{plotFun}
\alias{plotFun}
\title{Plot Function}
\usage{
plotFun(df, x, y)
}
\arguments{
\item{df}{The data frame containing the variables to be plotted.}

\item{x}{The name of the variable to be plotted on the x-axis.}

\item{y}{The name of the variable to be plotted on the y-axis.}
}
\value{
A ggplot object representing the scatter plot.
}
\description{
This function creates a scatter plot using ggplot2.
}
